+++ lib/solargraph/language_server/host.rb
@@ -2,6 +2,7 @@
 
 require 'observer'
 require 'set'
+require 'securerandom'
 
 module Solargraph
   module LanguageServer
@@ -277,6 +278,7 @@ module Solargraph
         begin
           lib = Solargraph::Library.load(path, name)
           libraries.push lib
+          async_library_map lib
         rescue WorkspaceTooLargeError => e
           send_notification 'window/showMessage', {
             'type' => Solargraph::LanguageServer::MessageTypes::WARNING,
@@ -740,6 +742,72 @@ module Solargraph
 
       def prepare_rename?
         client_capabilities['rename'] && client_capabilities['rename']['prepareSupport']
+      end
+
+      def client_supports_progress?
+        client_capabilities['window'] && client_capabilities['window']['workDoneProgress']
+      end
+
+      # @param library [Library]
+      # @return [void]
+      def async_library_map library
+        return if library.mapped?
+        Thread.new do
+          if client_supports_progress?
+            uuid = SecureRandom.uuid
+            send_request 'window/workDoneProgress/create', {
+              token: uuid
+            } do |response|
+              do_async_library_map library, response.nil? ? uuid : nil
+            end
+          else
+            do_async_library_map library
+          end
+        end
+      end
+
+      def do_async_library_map library, uuid = nil
+        total = library.workspace.sources.length
+        if uuid
+          send_notification '$/progress', {
+            token: uuid,
+            value: {
+              kind: 'begin',
+              title: "Mapping workspace",
+              message: "0/#{total} files",
+              cancellable: false,
+              percentage: 0
+            }
+          }
+        end
+        pct = 0
+        mod = 10
+        while library.next_map
+          cur = ((library.source_map_hash.keys.length.to_f / total.to_f) * 100).to_i
+          if cur > pct && cur % mod == 0
+            pct = cur
+            if uuid
+              send_notification '$/progress', {
+                token: uuid,
+                value: {
+                  kind: 'report',
+                  cancellable: false,
+                  message: "#{library.source_map_hash.keys.length}/#{total} files",
+                  percentage: pct
+                }
+              }
+            end
+          end
+        end
+        if uuid
+          send_notification '$/progress', {
+            token: uuid,
+            value: {
+              kind: 'end',
+              message: 'Mapping complete'
+            }
+          }
+        end
       end
     end
   end