+++ lib/solargraph/language_server/message/text_document/formatting.rb
@@ -19,10 +19,12 @@ module Solargraph
             require_rubocop(config['version'])
             options, paths = RuboCop::Options.new.parse(args)
             options[:stdin] = original
-            redirect_stdout do
+            corrections = redirect_stdout do
               RuboCop::Runner.new(options, RuboCop::ConfigStore.new).run(paths)
             end
             result = options[:stdin]
+
+            log_corrections(corrections)
 
             format original, result
           rescue RuboCop::ValidationError, RuboCop::ConfigNotFoundError => e
@@ -31,6 +33,17 @@ module Solargraph
 
           private
 
+          def log_corrections(corrections)
+            corrections = corrections&.strip
+            return if corrections&.empty?
+
+            Solargraph.logger.info('Formatting result:')
+            corrections.each_line do |line|
+              next if line.strip.empty?
+              Solargraph.logger.info(line.strip)
+            end
+          end
+
           def config_for(file_uri)
             conf = host.formatter_config(file_uri)
             return {} unless conf.is_a?(Hash)
@@ -38,7 +51,8 @@ module Solargraph
             conf['rubocop'] || {}
           end
 
-          def cli_args file, config
+          def cli_args file_uri, config
+            file = UriHelpers.uri_to_file(file_uri)
             args = [
               config['cops'] == 'all' ? '--auto-correct-all' : '--auto-correct',
               '--cache', 'false',