+++ lib/solargraph/source_map/mapper.rb
@@ -71,11 +71,6 @@ module Solargraph
           pos = Solargraph::Position.new(comment_position.line + line_num - 1, comment_position.column)
           process_directive(source_position, pos, d)
           last_line = line_num + 1
-          # @todo The below call assumes the topmost comment line. The above
-          #   process occasionally emits incorrect comment positions due to
-          #   blank lines in comment blocks, but at least it processes all the
-          #   directives.
-          # process_directive(source_position, comment_position, d)
         end
       end