+++ spec/library_spec.rb
@@ -416,4 +416,38 @@ describe Solargraph::Library do
     bar_alias = pins.select { |pin| pin.name == 'bar_alias' }.first
     expect(bar_alias.symbol_kind).to eq(Solargraph::LanguageServer::SymbolKinds::METHOD)
   end
+
+  context 'unsynchronized' do
+    let(:library) { Solargraph::Library.load File.absolute_path(File.join('spec', 'fixtures', 'workspace')) }
+    let(:good_file) { File.join(library.workspace.directory, 'lib', 'thing.rb') }
+    let(:bad_file) { File.join(library.workspace.directory, 'lib', 'not_a_thing.rb') }
+
+    describe 'Library#completions_at' do
+      it 'gracefully handles unmapped sources' do
+        expect {
+          library.completions_at(good_file, 0, 0)
+        }.not_to raise_error
+      end
+
+      it 'raises errors for nonexistent sources' do
+        expect {
+          library.completions_at(bad_file, 0, 0)
+        }.to raise_error(Solargraph::FileNotFoundError)
+      end
+    end
+
+    describe 'Library#definitions_at' do
+      it 'gracefully handles unmapped sources' do
+        expect {
+          library.definitions_at(good_file, 0, 0)
+        }.not_to raise_error
+      end
+
+      it 'raises errors for nonexistent sources' do
+        expect {
+          library.definitions_at(bad_file, 0, 0)
+        }.to raise_error(Solargraph::FileNotFoundError)
+      end
+    end
+  end
 end