+++ spec/source/source_chainer_spec.rb
@@ -247,4 +247,18 @@ describe Solargraph::Source::SourceChainer do
     chain = Solargraph::Source::SourceChainer.chain(source, Solargraph::Position.new(5, 13))
     expect(chain.links.last.word).to eq('Outer::Inner')
   end
+
+  it 'detects whole constant with cursor at double colon' do
+    source = Solargraph::Source.load_string(%(
+      class Outer
+        class Inner1
+          class Inner2
+          end
+        end
+      end
+      Outer::Inner1::Inner2.new
+    ), 'test.rb')
+    chain = Solargraph::Source::SourceChainer.chain(source, Solargraph::Position.new(7, 21))
+    expect(chain.links.last.word).to eq('Outer::Inner1::Inner2')
+  end
 end